+++ Runtime/Tweens/ValueTweens.cs
@@ -8,6 +8,7 @@ public static class ValueTweens {
     
     public static Tween<float> TweenValue(this Object obj, float from, float to) => obj.Tween(Value, to).From(from);
     public static Tween<float> TweenValue(this Object obj, float from, float to, float duration) => obj.Tween(Value, to).From(from).SetDuration(duration);
+    public static Tween<float> TweenValue01(this Object obj, float duration = 0) => obj.TweenValue(0, 1).SetDuration(duration);
     
     public static Tween<float> Delay(this Object obj, float duration) => obj.Tween(Value, 0).SetDuration(duration);
     public static Tween<float> DelayedCall(this Object obj, float duration, Action callback) => obj.Delay(duration).OnComplete(callback);